run lint and fixes
