+++ libs/elasticsearch/tests/unit_tests/test_imports.py
@@ -2,7 +2,7 @@ from langchain_elasticsearch import __all__
 
 EXPECTED_ALL = [
     "ElasticsearchCache",
-    "ElasticsearchStoreEmbeddings",
+    "ElasticsearchEmbeddingsCache",
     "ElasticsearchChatMessageHistory",
     "ElasticsearchEmbeddings",
     "ElasticsearchRetriever",
